# spec/openapi-edukate-backend.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:5810
security:
- cookieAuth: []
...
paths:
...
  /api/v1/checker/supervisor:
    post:
      tags:
      - Checker
      summary: Request supervisor check
      description: Requests a supervisor check for the provided submission (not implemented
        yet)
      operationId: supervisorCheck
      parameters:
      - name: id
        in: query
        description: Submission identifier
        required: true
        schema:
          type: string
      responses:
        "501":
          description: Not implemented
      security:
      - cookieAuth: []
  /api/v1/checker/self:
    post:
      tags:
      - Checker
      summary: Mark submission as self-checked
      description: Creates a self-check result for the provided submission
      operationId: selfCheck
      parameters:
      - name: id
        in: query
        description: Submission identifier
        required: true
        schema:
          type: string
      responses:
        "202":
          description: Self-check accepted
        "401":
          description: Unauthorized
        "403":
          description: Access denied
        "404":
          description: Submission not found
      security:
      - cookieAuth: []
  /api/v1/checker/ai:
    post:
      tags:
      - Checker
      summary: Schedule AI check
      description: Schedules an asynchronous AI check for the provided submission
      operationId: aiCheck
      parameters:
      - name: id
        in: query
        description: Submission identifier
        required: true
        schema:
          type: string
      responses:
        "202":
          description: Check scheduled
        "401":
          description: Unauthorized
        "403":
          description: Access denied
        "404":
          description: Submission not found
      security:
      - cookieAuth: []
...
  /api/v1/checker/submissions/{submissionId}:
    get:
      tags:
      - Checker
      summary: Get check results for submission
      description: Retrieves all check results (lightweight info) for a submission
        owned by the user
      operationId: getCheckResultsBySubmissionId
      parameters:
      - name: submissionId
        in: path
        description: Submission identifier
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successfully retrieved check results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CheckResultInfo"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CheckResultInfo"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CheckResultInfo"
      security:
      - cookieAuth: []
  /api/v1/checker/by-id/{id}:
    get:
      tags:
      - Checker
      summary: Get check result by id
      description: Retrieves a single check result by its identifier
      operationId: getCheckResultById
      parameters:
      - name: id
        in: path
        description: Check result identifier
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successfully retrieved check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckResultDto"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckResultDto"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckResultDto"
        "404":
          description: Check result not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckResultDto"
      security:
      - cookieAuth: []
...
  /api/v1/answers/{bookSlug}/{code}:
    get:
      tags:
      - Answers
      summary: Get answer by problem key
      description: Retrieves an answer for a specific problem
      operationId: getAnswerByProblemKey
      parameters:
      - name: bookSlug
        in: path
        description: Book slug
        required: true
        schema:
          type: string
      - name: code
        in: path
        description: Problem code
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successfully retrieved answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnswerDto"
        "404":
          description: Answer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnswerDto"
      security: []
components:
  schemas:
...
    CheckResultInfo:
      type: object
      properties:
        id:
          type: integer
          format: int64
        status:
          type: string
          enum:
          - SUCCESS
          - MISTAKE
          - INTERNAL_ERROR
        trustLevel:
          type: number
          format: float
        createdAt:
          type: string
          format: date-time
      required:
      - createdAt
      - id
      - status
      - trustLevel
    CheckResultDto:
      type: object
      properties:
        status:
          type: string
          enum:
          - SUCCESS
          - MISTAKE
          - INTERNAL_ERROR
        trustLevel:
          type: number
          format: float
        errorType:
          type: string
          enum:
          - ALGEBRAIC
          - CONCEPTUAL
          - INCOMPLETE
          - NONE
          - NUMERICAL
          - UNCLEAR
          - UNIT
        explanation:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
      - createdAt
      - errorType
      - explanation
      - status
      - trustLevel
...
    AnswerDto:
      type: object
      properties:
        text:
          type: string
        notes:
          type: string
        images:
          type: array
          items:
            type: string
      required:
      - images
      - text
  securitySchemes:
    cookieAuth:
      type: apiKey
      description: JWT authentication cookie
      name: X-Auth
      in: cookie